requirements.txt
streamlit==0.89.0
pandas==1.3.3
numpy==1.19.5
matplotlib==3.4.3
networkx==2.6.3
scikit-learn==1.0
mecab-python3==1.0.4
